# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GrubPac ERP API
  version: 0.1.0
  description: |
    Modular monolith ERP platform — foundation HTTP API for GrubPac ERP (GrubFleet-ERP).

    All routes are served under the global prefix `/api/v1` (override via backend `API_PREFIX`).

    Live Swagger UI when the backend is running: `/api/v1/docs`.

    Authentication: login, refresh, logout, and `/auth/me` (Bearer JWT access token).
    Org-scoped admin routes require Bearer JWT plus organization context (`X-Organization-Id` and/or `organizationId` query/body). Module access levels are NONE / VIEW / MANAGE / FULL; permission keys use dot notation (e.g. `administration.view`, `administration.create`). MANAGE expands to view plus CRUD keys; FULL adds legacy `{module}.manage`. Admin mutations accept `{module}.manage` or the matching CRUD key. TSV Edit/Approve/Admin types deferred — see auth-rbac-plan.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development (default backend port 4000)
  - url: https://api-staging.example.com/api/v1
    description: Staging (placeholder — set in GitHub Environment `staging`)
  - url: https://api-preprod.example.com/api/v1
    description: Pre-production (placeholder — set in GitHub Environment `pre-production`)
  - url: https://api.example.com/api/v1
    description: Production (placeholder — set in GitHub Environment `production`)
...
paths:
...
  /asset-register/asset-classes:
    get:
      tags: [asset-register]
      summary: List asset classes (paginated)
      description: >-
        Default sort: createdAt descending (newest first), then id descending.
        Search matches class name or generated code. inFleetCount is active fleet register vehicles per class; availableCount is active vehicles with operationalStatus available (batched on list page).
      operationId: listAssetRegisterAssetClasses
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationIdQuery'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
      responses:
        '200':
          description: Paginated asset class rows
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags: [asset-register]
      summary: Create asset class (code generated server-side)
      operationId: createAssetRegisterAssetClass
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRegisterAssetClassRequest'
      responses:
        '201':
          description: Asset class created with generated code
        '400':
          description: Validation error or client-supplied code rejected
        '403':
          description: Forbidden
  /asset-register/asset-classes/{id}:
    get:
      tags: [asset-register]
      summary: Asset class detail
      operationId: getAssetRegisterAssetClass
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationIdQuery'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Asset class detail (code read-only)
        '404':
          description: Not found
    patch:
      tags: [asset-register]
      summary: Update asset class
      operationId: updateAssetRegisterAssetClass
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationIdQuery'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssetRegisterAssetClassRequest'
      responses:
        '200':
          description: Updated asset class
        '400':
          description: Validation error or inactive asset class cannot be edited
  /asset-register/asset-classes/{id}/status:
    patch:
      tags: [asset-register]
      summary: Activate or deactivate asset class (deactivate requires reason)
      operationId: updateAssetRegisterAssetClassStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationIdQuery'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [activate, deactivate]
                reason:
                  type: string
                  maxLength: 500
                  description: Required when action is deactivate
      responses:
        '200':
          description: Status updated
        '400':
          description: Missing deactivate reason or invalid transition
  /asset-register/asset-classes/{classId}/masters:
    get:
      tags: [asset-register]
      summary: Catalog — active asset masters for an active asset class
      operationId: listAssetRegisterAssetMastersByClass
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationIdQuery'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated active masters for fleet register dropdown
        '400':
          description: Asset class is inactive
        '404':
          description: Asset class not found
...
  /asset-register/vehicles:
    get:
      tags: [asset-register]
      summary: List fleet register vehicles (paginated)
      description: >-
        fleetCode, registrationNumber, asset class name, odometer, operationalStatus.
        Filter by assetClassId, operationalStatus, lifecycle status (active/inactive), or search.
      operationId: listAssetRegisterVehicles
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationIdQuery'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: assetClassId
          in: query
          schema:
            type: string
            format: uuid
        - name: operationalStatus
          in: query
          schema:
            type: string
            enum: [available, leased, workshop, sold, retired]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated fleet register vehicles
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags: [asset-register]
      summary: Register a vehicle (fleetCode server-generated)
      operationId: createAssetRegisterVehicle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRegisterVehicleRequest'
      responses:
        '201':
          description: Created vehicle with VH- sequence fleetCode
        '400':
          description: Validation error or inactive class/master
        '403':
          description: Forbidden
  /asset-register/vehicles/{id}:
    get:
      tags: [asset-register]
      summary: Fleet register vehicle detail
      operationId: getAssetRegisterVehicle
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationIdQuery'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vehicle with class and master names
        '404':
          description: Not found
    patch:
      tags: [asset-register]
      summary: Update fleet register vehicle
      operationId: updateAssetRegisterVehicle
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationIdQuery'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssetRegisterVehicleRequest'
      responses:
        '200':
          description: Updated vehicle
        '400':
          description: Inactive vehicle or validation error
  /asset-register/vehicles/{id}/status:
    patch:
      tags: [asset-register]
      summary: Activate or deactivate fleet register vehicle
      operationId: updateAssetRegisterVehicleStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationIdQuery'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [activate, deactivate]
                reason:
                  type: string
                  maxLength: 500
                  description: Required when action is deactivate
      responses:
        '200':
          description: Status updated
        '400':
          description: Missing deactivate reason or invalid transition
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    OrganizationIdQuery:
      name: organizationId
      in: query
      required: true
      description: >-
        Organization UUID for org-scoped routes. Alternatively (or additionally)
        send `X-Organization-Id`; the header is preferred when both are present.
        POST bodies may include `organizationId` instead of query/header where documented.
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateAssetRegisterAssetClassRequest:
      type: object
      required:
        - organizationId
        - name
        - vehicleType
        - fuelType
        - fuelTankCapacity
        - ratedLoadFrom
        - ratedLoadTo
      properties:
        organizationId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        vehicleType:
          type: string
          enum: ['2W', '3W', '4W']
        fuelType:
          type: string
          maxLength: 64
        mileageFrom:
          type: number
          minimum: 0
        mileageTo:
          type: number
          minimum: 0
        mileageUnit:
          type: string
          maxLength: 32
        fuelTankCapacity:
          type: number
          minimum: 0
        ratedLoadFrom:
          type: number
          minimum: 0
        ratedLoadTo:
          type: number
          minimum: 0
        defaultIntakeChecklist:
          type: string
          maxLength: 2000
      additionalProperties: false
      description: Class code is generated server-side; do not send code on create.
    UpdateAssetRegisterAssetClassRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        vehicleType:
          type: string
          enum: ['2W', '3W', '4W']
        fuelType:
          type: string
          maxLength: 64
        mileageFrom:
          type: number
          minimum: 0
        mileageTo:
          type: number
          minimum: 0
        mileageUnit:
          type: string
          maxLength: 32
        fuelTankCapacity:
          type: number
          minimum: 0
        ratedLoadFrom:
          type: number
          minimum: 0
        ratedLoadTo:
          type: number
          minimum: 0
        defaultIntakeChecklist:
          type: string
          maxLength: 2000
      additionalProperties: false
...
    CreateAssetRegisterVehicleRequest:
      type: object
      required:
        - organizationId
        - assetClassId
        - assetMasterId
        - registrationNumber
        - chassisNumber
        - modelYear
        - odometer
        - registrationStartDate
        - registrationEndDate
        - insuranceStartDate
        - insuranceEndDate
        - insurancePremium
        - warrantyStartDate
        - warrantyEndDate
      properties:
        organizationId:
          type: string
          format: uuid
        assetClassId:
          type: string
          format: uuid
        assetMasterId:
          type: string
          format: uuid
        registrationNumber:
          type: string
          maxLength: 32
        chassisNumber:
          type: string
          maxLength: 64
        modelYear:
          type: integer
          minimum: 1900
          maximum: 2100
        odometer:
          type: integer
          minimum: 0
          maximum: 9999999
        registrationStartDate:
          type: string
          format: date
        registrationEndDate:
          type: string
          format: date
        insuranceStartDate:
          type: string
          format: date
        insuranceEndDate:
          type: string
          format: date
        insurancePremium:
          type: number
          minimum: 0
        warrantyStartDate:
          type: string
          format: date
        warrantyEndDate:
          type: string
          format: date
        specialNotes:
          type: string
          maxLength: 2000
        purchaseInvoiceId:
          type: string
          format: uuid
          nullable: true
      additionalProperties: false
      description: >-
        fleetCode is server-generated (VH-1001 sequence per org). operationalStatus defaults to available.
        purchaseInvoiceId accepted without Finance FK validation in v1.
    UpdateAssetRegisterVehicleRequest:
      type: object
      properties:
        assetClassId:
          type: string
          format: uuid
        assetMasterId:
          type: string
          format: uuid
        registrationNumber:
          type: string
          maxLength: 32
        chassisNumber:
          type: string
          maxLength: 64
        modelYear:
          type: integer
          minimum: 1900
          maximum: 2100
        odometer:
          type: integer
          minimum: 0
          maximum: 9999999
        registrationStartDate:
          type: string
          format: date
        registrationEndDate:
          type: string
          format: date
        insuranceStartDate:
          type: string
          format: date
        insuranceEndDate:
          type: string
          format: date
        insurancePremium:
          type: number
          minimum: 0
        warrantyStartDate:
          type: string
          format: date
        warrantyEndDate:
          type: string
          format: date
        specialNotes:
          type: string
          maxLength: 2000
        purchaseInvoiceId:
          type: string
          format: uuid
          nullable: true
      additionalProperties: false
      description: operationalStatus is not client-writable in v1 (system/assignment modules later).
...